requirements.txt
appdirs==1.4.4
asgiref==3.5.0
attrs==21.2.0
black==19.10b0
click==8.0.3
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.10.0
django-heroku==0.3.1
django-nose==1.4.7
djangorestframework==3.13.1
factory-boy==2.12.0
Faker==9.5.2
flake8==3.8.3
gunicorn==20.1.0
isort==5.2.1
mccabe==0.6.1
nose==1.3.7
pathspec==0.9.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==2.1.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.14.0
pytz==2022.1
regex==2021.10.21
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
whitenoise==6.0.0
